tools/atlas_baker/requirements.txt
Pillow>=10.0
rectpack>=0.2.2
